backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=24.2
Pillow>=10.0
qrcode[pil]>=7.4
drf-spectacular>=0.27
python-dotenv>=1.0
django-cors-headers>=4.3

